Minor refactor of old stuff
